Переносит логику показа сообщений в отдельный модуль
