+++ js/validation.js
@@ -1,4 +1,5 @@
 import { sendData} from './api';
+import { addSuccessMessage, addErrorMessage } from './message';
 
 const MAX_HASHTAGS = 5;
 const COMMENT_MAX_LENGTH = 140;
@@ -13,87 +14,6 @@ const hashtagInput = imageUploadForm.querySelector('.text__hashtags');
 const commentInput = imageUploadForm.querySelector('.text__description');
 const submitButton = imageUploadForm.querySelector('#upload-submit');
 const hashtagPattern = /^#[a-zA-Zа-яА-ЯёЁ0-9]{1,19}$/i;
-const errorDataMessageTemplate = document.querySelector('#data-error').content;
-const successMessageTemplate = document.querySelector('#success').content;
-const errorMessageTemplate = document.querySelector('#error').content;
-
-const messageFragment = document.createDocumentFragment();
-
-const addSuccessMessage = () => {
-  const successMessage = successMessageTemplate.cloneNode(true);
-  messageFragment.append(successMessage);
-  document.body.append(messageFragment);
-
-  const successButton = document.querySelector('.success__button');
-  const successElement = document.querySelector('.success');
-  const successElementModal = successElement.querySelector('.success__inner');
-
-  const onClickButton = () => {
-    successElement.remove();
-    successButton.removeEventListener('click', onClickButton);
-  };
-
-  const onEscapeKeydown = (evt) => {
-    if (evt.key === 'Escape') {
-      successElement.remove();
-      document.removeEventListener('keydown', onEscapeKeydown);
-    }
-  };
-
-  const onClickOverlay = (evt) => {
-    if(successElement.className === 'success' && !successElementModal.contains(evt.target)) {
-      successElement.remove();
-      document.removeEventListener('click', onClickOverlay);
-    }
-  };
-
-  successButton.addEventListener('click', onClickButton);
-  document.addEventListener('keydown', onEscapeKeydown);
-  document.addEventListener('click', onClickOverlay);
-};
-
-const addErrorDataMessage = () => {
-  const errorMessage = errorDataMessageTemplate.cloneNode(true);
-  messageFragment.append(errorMessage);
-  document.body.append(messageFragment);
-  setTimeout(() => {
-    const errorElement = document.querySelector('.data-error');
-    errorElement.remove();
-  }, 5000);
-};
-
-const addErrorMessage = () => {
-  const errorMessage = errorMessageTemplate.cloneNode(true);
-  messageFragment.append(errorMessage);
-  document.body.append(messageFragment);
-
-  const errorButton = document.querySelector('.error__button');
-  const errorElement = document.querySelector('.error');
-
-  const errorInner = errorElement.querySelector('.error__inner');
-
-  const onClickButton = () => {
-    errorElement.remove();
-    errorButton.removeEventListener('click', onClickButton);
-  };
-  const onEscapeKeydown = (evt) => {
-    if (evt.key === 'Escape') {
-      errorElement.remove();
-      document.removeEventListener('keydown', onEscapeKeydown);
-    }
-  };
-  const onClickOverlay = (evt) => {
-    if (!errorInner.contains(evt.target)) {
-      evt.stopPropagation();
-      errorElement.remove();
-      document.removeEventListener('click', onClickOverlay);
-    }
-  };
-
-  errorButton.addEventListener('click', onClickButton);
-  document.addEventListener('keydown', onEscapeKeydown);
-  document.addEventListener('click', onClickOverlay);
-};
 
 const isInputFocused = () => hashtagInput === document.activeElement || commentInput === document.activeElement;
 const imageUploadValidator = new Pristine(imageUploadForm, {
@@ -183,4 +103,4 @@ const setImageFormSubmit = (onSuccess) => {
 imageUploadValidator.addValidator(hashtagInput, validateHashtags, getHashtagErrorMessage);
 imageUploadValidator.addValidator(commentInput, validateComment, getCommentErrorMessage);
 
-export { imageUploadValidator, setImageFormSubmit, isInputFocused, addErrorDataMessage, addErrorMessage };
+export { imageUploadValidator, setImageFormSubmit, isInputFocused };
